Update와 Delete에 대한 권한 설정

+++ posts/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .forms import PostModelForm
 from .models import Post
 
@@ -34,13 +34,17 @@ def list(request):
     
     
 def delete(request, post_id):
-    post = Post.objects.get(id=post_id)
+    post = get_object_or_404(Post, id=post_id)
+    if post.user != request.user:
+        return redirect('posts:list')
     post.delete()
     return redirect('posts:list')
     
     
 def update(request, post_id):
-    post = Post.objects.get(id=post_id)
+    post = get_object_or_404(Post, id=post_id)
+    if post.user != request.user:
+        return redirect('posts:list')
     if request.method == 'POST':
         # 수정내용 DB에 반영
         form = PostModelForm(request.POST, instance=post)
